Add cui::status_pane namespace
